Hide extend_units
